requirements.txt
django==3.1.3
numpy==1.19.2
opencv-python
pandas==1.1.3
pillow==8.0.1
pip==20.2.4
psycopg2-binary>=2.8
torch==1.7.0
